+++ Forms/DownloadForm.vb
@@ -13,6 +13,7 @@ Public Class DownloadForm
         AddHandler WebClient.DownloadProgressChanged, AddressOf ProgressChanged
         AddHandler WebClient.DownloadFileCompleted, AddressOf Completed
         Dim filename = url.FileName
+        WebClient.Headers.Add("user-agent", "Mozilla/5.0 (Windows NT 10.0; Win64; x64) AppleWebKit/537.36 (KHTML, like Gecko) Chrome/74.0.3729.169 Safari/537.36")
         WebClient.DownloadFileAsync(New Uri(url), path)
     End Sub