+++ Changelog.md
@@ -2,8 +2,7 @@
 2.1.4.8 Beta (2020-09-??)
 ============
 
-- The Apps dialog supports drag & drop for zip/7z files,
-  mostly usefull for the tool update maintainer (stax76)
+- The Apps dialog supports drag & drop for zip/7z files (stax76)
 - qaac 2.71 (long path support)
 - chapterEditor 1.24
 - mkvtoolnix 51