+++ Changelog.md
@@ -4,6 +4,7 @@
 
 - The Apps dialog supports drag & drop for zip/7z files (stax76)
 - aomenc fixes and improvements (stax76, [#357](/../../issues/357))
+- GUI options for ffmpeg -probesize -analyzeduration (stax76, [#354](/../../issues/354))
 - qaac 2.71 (long path support)
 - chapterEditor 1.24
 - mkvtoolnix 51